[Console] Send questions set (#26256)
- Related to: #25819

## Summary

- Send questions set in chat
- Message update when questions set is updated

## Important notes for @gregoirenovel ##

### 1 - Questions-set draft ###
Because of how the console is currently implemented, questions-set are
not send immediately but are put in draft in the message composer.
From here, the questions set can be deleted. It make also sense to be
able to read it. At the moment, I just reused what I just did and allow
to re open the questions set composer, but it gives weird wording.
Another solution could be to open the questions-set in read only mode
(the same way provider will consult questions-set answers).

Let me know what do you prefer:
- [x] Allow edit (new wording needed)
- [ ] Allow read only
- [ ] Something else

### 2- Design ###
I reused what we have for `action request` type message to avoid to
redesign close but different items for questions-set. Let me know if it
is ok for v0.
- [x] Ok
- [ ] Not ok 

## Additional context


https://user-images.githubusercontent.com/7696995/233392806-3ccba553-7f4e-4c63-ae90-60f025b06d0f.mov